+++ Sources/BraintreeCore/Analytics/BTAnalyticsService.swift
@@ -4,18 +4,36 @@ class BTAnalyticsService: Equatable {
 
     // MARK: - Internal Properties
 
-    /// The HTTP client for communication with the analytics service endpoint. Exposed for testing.
-    var http: BTHTTP?
-    
     /// The FPTI URL to post all analytic events.
     static let url = URL(string: "https://api-m.paypal.com")!
 
-    private let apiClient: BTAPIClient
+    /// The HTTP client for communication with the analytics service endpoint. Exposed for testing.
+    var http: BTHTTP?
 
+    /// Exposed for testing only
+    var shouldBypassTimerQueue = false
+
+    // MARK: - Private Properties
+
+    private static let events = BTAnalyticsEventsStorage()
+
+    private let apiClient: BTAPIClient
+    /// Amount of time, in seconds, between batch API requests sent to FPTI
+    private let timerInterval = 15
+
+    private static var timer: DispatchSourceTimer?
+    
     // MARK: - Initializer
 
     init(apiClient: BTAPIClient) {
         self.apiClient = apiClient
+    }
+
+    // MARK: - Deinit
+
+    deinit {
+        BTAnalyticsService.timer?.cancel()
+        BTAnalyticsService.timer = nil
     }
 
     // MARK: - Internal Methods
@@ -23,24 +41,35 @@ class BTAnalyticsService: Equatable {
     /// Sends analytics event to https://api.paypal.com/v1/tracking/batch/events/ via a background task.
     /// - Parameters:
     ///   - eventName: Name of analytic event.
+    ///   - isVaultRequest: Optional. If the Venmo or PayPal request is being vaulted.
     ///   - correlationID: Optional. CorrelationID associated with the checkout session.
+    ///   - endpoint: Optional. The endpoint of the API request send during networking requests.
+    ///   - endTime: Optional. The end time of the roundtrip networking request.
     ///   - errorDescription: Optional. Full error description returned to merchant.
     ///   - linkType: Optional. The type of link the SDK will be handling, currently deeplink or universal.
     ///   - payPalContextID: Optional. PayPal Context ID associated with the checkout session.
+    ///   - startTime: Optional. The start time of the networking request.
     func sendAnalyticsEvent(
         _ eventName: String,
+        isVaultRequest: Bool? = nil,
         correlationID: String? = nil,
+        endpoint: String? = nil,
+        endTime: Int? = nil,
         errorDescription: String? = nil,
         linkType: String? = nil,
-        payPalContextID: String? = nil
+        payPalContextID: String? = nil,
+        startTime: Int? = nil
     ) {
         Task(priority: .background) {
             await performEventRequest(
                 eventName,
                 correlationID: correlationID,
+                endpoint: endpoint,
+                endTime: endTime,
                 errorDescription: errorDescription,
                 linkType: linkType,
-                payPalContextID: payPalContextID
+                payPalContextID: payPalContextID,
+                startTime: startTime
             )
         }
     }
@@ -48,26 +77,34 @@ class BTAnalyticsService: Equatable {
     /// Exposed to be able to execute this function synchronously in unit tests
     func performEventRequest(
         _ eventName: String,
+        isVaultRequest: Bool? = nil,
         correlationID: String? = nil,
+        endpoint: String? = nil,
+        endTime: Int? = nil,
         errorDescription: String? = nil,
         linkType: String? = nil,
-        payPalContextID: String? = nil
+        payPalContextID: String? = nil,
+        startTime: Int? = nil
     ) async {
-        let timestampInMilliseconds = UInt64(Date().timeIntervalSince1970 * 1000)
+        let timestampInMilliseconds = Date().utcTimestampMilliseconds
         let event = FPTIBatchData.Event(
             correlationID: correlationID,
+            endpoint: endpoint,
+            endTime: endTime,
             errorDescription: errorDescription,
             eventName: eventName,
+            isVaultRequest: isVaultRequest,
             linkType: linkType,
             payPalContextID: payPalContextID,
+            startTime: startTime,
             timestamp: String(timestampInMilliseconds)
         )
-                
-        apiClient.fetchOrReturnRemoteConfiguration { configuration, error in
-            guard let configuration, error == nil else {
-                return
-            }
 
+        await BTAnalyticsService.events.append(event)
+        
+        do {
+            let configuration = try await apiClient.fetchConfiguration()
+            
             // TODO: - Refactor to make HTTP non-optional property and instantiate in init()
             if self.http == nil {
                 if let clientToken = self.apiClient.clientToken {
@@ -83,16 +120,45 @@ class BTAnalyticsService: Equatable {
             if let http = self.http, http.baseURL.absoluteString == "test://do-not-send.url" {
                 return
             }
+            
+            if shouldBypassTimerQueue {
+                await self.sendQueuedAnalyticsEvents(configuration: configuration)
+                return
+            }
 
-            let postParameters = self.createAnalyticsEvent(config: configuration, sessionID: self.apiClient.metadata.sessionID, event: event)
-            self.http?.post("v1/tracking/batch/events", parameters: postParameters) { _, _, _ in }
+            if BTAnalyticsService.timer == nil {
+                BTAnalyticsService.timer = DispatchSource.makeTimerSource(queue: self.http?.dispatchQueue)
+                BTAnalyticsService.timer?.schedule(
+                    deadline: .now() + .seconds(self.timerInterval),
+                    repeating: .seconds(self.timerInterval),
+                    leeway: .seconds(1)
+                )
+
+                BTAnalyticsService.timer?.setEventHandler {
+                    Task {
+                        await self.sendQueuedAnalyticsEvents(configuration: configuration)
+                    }
+                }
+
+                BTAnalyticsService.timer?.resume()
+            }
+        } catch {
+            return
         }
     }
 
     // MARK: - Helpers
 
+    func sendQueuedAnalyticsEvents(configuration: BTConfiguration) async {
+        if await !BTAnalyticsService.events.isEmpty {
+            let postParameters = await createAnalyticsEvent(config: configuration, sessionID: apiClient.metadata.sessionID, events: BTAnalyticsService.events.allValues)
+            http?.post("v1/tracking/batch/events", parameters: postParameters) { _, _, _ in }
+            await BTAnalyticsService.events.removeAll()
+        }
+    }
+
     /// Constructs POST params to be sent to FPTI
-    func createAnalyticsEvent(config: BTConfiguration, sessionID: String, event: FPTIBatchData.Event) -> Codable {
+    func createAnalyticsEvent(config: BTConfiguration, sessionID: String, events: [FPTIBatchData.Event]) -> Codable {
         let batchMetadata = FPTIBatchData.Metadata(
             authorizationFingerprint: apiClient.clientToken?.authorizationFingerprint,
             environment: config.fptiEnvironment,
@@ -102,7 +168,7 @@ class BTAnalyticsService: Equatable {
             tokenizationKey: apiClient.tokenizationKey
         )
         
-        return FPTIBatchData(metadata: batchMetadata, events: [event])
+        return FPTIBatchData(metadata: batchMetadata, events: events)
     }
 
     // MARK: Equitable Protocol Conformance