+++ Sources/BraintreeCore/BTGraphQLHTTP.swift
@@ -10,20 +10,12 @@ class BTGraphQLHTTP: BTHTTP {
 
     // MARK: - Overrides
 
-    override func get(_ path: String, parameters: Encodable? = nil, shouldCache: Bool = false, completion: @escaping RequestCompletion) {
+    override func get(_ path: String, parameters: Encodable? = nil, completion: @escaping RequestCompletion) {
         NSException(name: exceptionName, reason: "GET is unsupported").raise()
     }
 
     override func post(_ path: String, parameters: [String: Any]? = nil, completion: @escaping RequestCompletion) {
         httpRequest(method: "POST", parameters: parameters, completion: completion)
-    }
-
-    override func put(_ path: String, parameters: [String: Any]? = nil, completion: @escaping RequestCompletion) {
-        NSException(name: exceptionName, reason: "PUT is unsupported").raise()
-    }
-
-    override func delete(_ path: String, parameters: [String: Any]? = nil, completion: @escaping RequestCompletion) {
-        NSException(name: exceptionName, reason: "DELETE is unsupported").raise()
     }
 
     // MARK: - Internal methods