+++ UnitTests/BraintreeCoreTests/BTGraphQLHTTP_Tests.swift
@@ -65,30 +65,6 @@ final class BTGraphQLHTTP_Tests: XCTestCase {
         }
     }
 
-    func testPUTRequests_areUnsupported() {
-        do {
-            try BTExceptionCatcher.catchException {
-                self.http?.put("") { _, _, _ in
-                    // no-op
-                }
-            }
-        } catch let error as NSError {
-            XCTAssertEqual(error.userInfo["ExceptionReason"] as! String, "PUT is unsupported")
-        }
-    }
-
-    func testDELETERequests_areUnsupported() {
-        do {
-            try BTExceptionCatcher.catchException {
-                self.http?.delete("") { _, _, _ in
-                    // no-op
-                }
-            }
-        } catch let error as NSError {
-            XCTAssertEqual(error.userInfo["ExceptionReason"] as! String, "DELETE is unsupported")
-        }
-    }
-
     // MARK: - POST requests
 
     func testPOSTRequests_sendsParametersInBody() {