+++ UnitTests/BraintreeTestShared/MockAPIClient.swift
@@ -9,6 +9,7 @@ public class MockAPIClient: BTAPIClient {
     public var lastGETPath = ""
     public var lastGETParameters = [:] as [String: Any]?
     public var lastGETAPIClientHTTPType: BTAPIClientHTTPService?
+    public var lastPostedVaultType = false
 
     public var postedAnalyticsEvents : [String] = []
     public var postedPayPalContextID: String? = nil
@@ -84,6 +85,7 @@ public class MockAPIClient: BTAPIClient {
 
     public override func sendAnalyticsEvent(
         _ name: String,
+        isVaultRequest: Bool? = nil,
         correlationID: String? = nil,
         errorDescription: String? = nil,
         linkType: String? = nil,