+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Braintree iOS SDK Release Notes
 
 ## unreleased
+* BraintreeCore
+  * Batch analytics events to FPTI
+
+## 6.18.2 (2024-05-15)
+* BraintreePayPal
+  * Send `start_time`, `end_time`, and `endpoint` to FPTI for tracking API request latency
+  * Send `isVaultRequest` to FPTI for tracking in Venmo and PayPal flows
+
+## 5.26.0 (2024-05-07)
+* Updated expiring pinned vendor SSL certificates
+
+## 6.18.1 (2024-05-06)
+* Remove throttle delay in accessing configuration, added in v5.9.0
+  * Move from URLCache to NSCache for configuration caching
+
+## 6.18.0 (2024-04-25)
 * BraintreePayPalNativeCheckout
   * Bump PayPalCheckout to version 1.3.0 with code signing & a privacy manifest file.
 * BraintreePayPalMessaging (BETA)